pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.7.0"
description = "Консольная утилита, определяющая разницу между двумя структурами данных."
authors = ["Anna Bystrova <annbystr@gmail.com>"]
readme = "README.md"
repository = "https://github.com/Aluwian/python-project-50"
packages = [{include = "gendiff"}]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Environment :: Console",
    "Topic :: Utilities"
]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.2.2"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"